extract ChatMessageContent - phase 6
